Product Service & validation
