--- src/main/java/com/javaguru/shoppinglist/service/validation/ProductDiscountValidationRule.java
@@ -1,15 +1,0 @@
-package com.javaguru.shoppinglist.service.validation;
-
-import com.javaguru.shoppinglist.domain.Product;
-
-import java.math.BigDecimal;
-
-public class ProductDiscountValidationRule {
-
-    public void validate(Product product) {
-        BigDecimal maxDiscount = new BigDecimal(80);
-        if (product.getDiscount().compareTo(BigDecimal.ZERO) < 0 || product.getDiscount().compareTo(maxDiscount) > 0) {
-            throw new ProductValidationException("Discount must be more than 0% and less than 80%");
-        }
-    }
-}
