+++ src/main/java/com/javaguru/shoppinglist/service/validation/ProductValidationRule.java
@@ -5,7 +5,7 @@ import com.javaguru.shoppinglist.domain.Product;
 public interface ProductValidationRule {
     void validate(Product product);
 
-    default void checkNotNull (Product product) {
+    default void checkNotNull(Product product) {
         if (product == null) {
             throw new ProductValidationException("Product must be not null");
         }
